pyproject.toml
[tool.poetry]
name = "mental-helth-ai"
version = "0.1.0"
description = "Projeto de Conclusão de Curso com o objetivo de desenvolver um chatbot especialista em saúde mental"
authors = ["Marco Antônio Martins Porto Netto"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
langchain = "^0.2.11"
pypdf2 = "^3.0.1"
langchain-community = "^0.2.10"
pypdf = "^4.3.1"
nltk = "^3.8.1"
rich = "^13.7.1"
fastapi = "^0.111.1"
faiss-cpu = "^1.8.0.post1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"
sentence-transformers = "^3.0.1"


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev mental_helth_ai/main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.5.5"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
